+++ collections/collections.go
@@ -328,6 +328,10 @@ type Tree interface {
 	// DoAll executes the passed function on all nodes.
 	DoAll(f func(value interface{}) error) error
 
+	// DoAllDeep executes the passed function on all nodes
+	// passing a deep list of values ordered top-down.
+	DoAllDeep(f func(values []interface{}) error) error
+
 	// Len returns the number of nodes of the tree.
 	Len() int
 
@@ -368,6 +372,10 @@ type StringTree interface {
 	// DoAll executes the passed function on all nodes.
 	DoAll(f func(value string) error) error
 
+	// DoAllDeep executes the passed function on all nodes
+	// passing a deep list of values ordered top-down.
+	DoAllDeep(f func(values []string) error) error
+
 	// Len returns the number of nodes of the tree.
 	Len() int
 
@@ -407,6 +415,10 @@ type KeyValueTree interface {
 
 	// DoAll executes the passed function on all nodes.
 	DoAll(f func(key string, value interface{}) error) error
+
+	// DoAllDeep executes the passed function on all nodes
+	// passing a deep list of keys ordered top-down.
+	DoAllDeep(f func(keys []string, value interface{}) error) error
 
 	// Len returns the number of nodes of the tree.
 	Len() int
@@ -451,6 +463,10 @@ type KeyStringValueTree interface {
 
 	// DoAll executes the passed function on all nodes.
 	DoAll(f func(key, value string) error) error
+
+	// DoAllDeep executes the passed function on all nodes
+	// passing a deep list of keys ordered top-down.
+	DoAllDeep(f func(keys []string, value string) error) error
 
 	// Len returns the number of nodes of the tree.
 	Len() int